Fixed the msvc release build and fixed some warnings in src/arch/win32/hardsid.c
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@19406 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/hardsid.c
@@ -127,7 +127,7 @@ void hardsid_reset(void)
         {
             for (chipno = 0; chipno < 4; chipno++)
             {
-                HardSID_AbortPlay(chipno);
+                HardSID_AbortPlay((BYTE)chipno);
             }
         }
     }
@@ -155,7 +155,7 @@ int hardsid_open(void)
     {
         for (chipno = 0; chipno < 4; chipno++)
         {
-            HardSID_Reset(chipno);
+            HardSID_Reset((BYTE)chipno);
         }
         lastaccess_clk = 0;
         lastaccess_ms = 0;
@@ -184,7 +184,7 @@ static void usb_hardsid_close(void)
     KillTimer(NULL, ftimer);
     for (chipno = 0; chipno < 4; chipno++)
     {
-        HardSID_AbortPlay(chipno);
+        HardSID_AbortPlay((BYTE)chipno);
     }
 }
